refactor(App): move `App` to containers/App/ ♻️
