+++ src/containers/App/App.jsx
@@ -1,7 +1,5 @@
 import React, { Component } from 'react';
 
-import 'src/styles/App.css';
-
 class App extends Component {
   render() {
     return (
